+++ R/plotWellReport.R
@@ -230,10 +230,12 @@ plotWellReportPPT <- function(csite, substances, locations, use_log_scale,
   dev.off()
   
   # Put into powerpoint slide.
-  AddPlotPPV2(mytemp, width, height) 
+  if (is.null(ppt_lst <- initPPT())) {
+    showNotification("Unable to initialize Powerpoint: package RDCOMClient might not be installed.", type = "error", duration = 10)
+    return(NULL)
+  }
   
+  addPlotPPT(mytemp, ppt_lst, width, height) 
+
   try(file.remove(mytemp))
-  
-  # This closes the handler
-  ppt <<- NULL
 }
